// src/enums/DiscussionEventType.ts
export enum DiscussionEventType {
    AGREED = 'agreed',
    DECLINED = 'declined',
    DETAILS_CHANGED = 'details_changed',
    INFORMATION_UPDATED = 'information_updated',
    REOPENED = 'reopened',
    CANCELLED = 'cancelled',
    REQUESTED = 'requested',
    NO_LONGER_VALID = 'no_longer_valid',
    LEGACY = 'legacy',
}